Cambio de double a decimal en el campo monto

+++ Liamell_Cruz_p1_Ap1/Models/Aportes.cs
@@ -20,7 +20,7 @@ public class Aportes
     public string Observacion { get; set; }
 
     [Required(ErrorMessage = "Campo Obligatorio")]
-    public double Monto { get; set; }
+    public decimal Monto { get; set; }
